refactored routes approach
