+++ public/index.html
@@ -36,7 +36,7 @@
     <!-- Open Graph Meta Tags (Facebook, LinkedIn, etc.) -->
   <meta property="og:title" content="Home | About | Rhythm Messengers">
   <meta property="og:description" content="Welcome to The Rhythm Messengers, the leading entertainment for events.">
-  <meta property="og:image" content="https://rhythmmessengers.com/media/RMELogo300-2-no-background.png">
+  <meta property="og:image" content="https://rhythmmessengers.com/media/RMELogo300-2-no-background.webp">
   <meta property="og:url" content="https://rhythmmessengers.com">
   <meta property="og:type" content="website">
   <meta property="og:site_name" content="The Rhythm Messengers">
@@ -50,12 +50,12 @@
   <meta name="theme-color" content="#000000"> <!-- Set the theme color for Android taskbar (Chrome on Android) -->
 
   <!-- Icon for Apple Touch (Icon when saved to Home Screen) -->
-  <link rel="apple-touch-icon" href="https://rhythmmessengers.com/media/RMELogo300-2-no-background.png">
+  <link rel="apple-touch-icon" href="https://rhythmmessengers.com/media/RMELogo300-2-no-background.webp">
 
   <!-- Icon for Android/General (Icon when saved to Home Screen) -->
-  <link rel="icon" href="https://rhythmmessengers.com/media/RMELogo300-2-no-background.png" type="image/x-icon">
+  <link rel="icon" href="https://rhythmmessengers.com/media/RMELogo300-2-no-background.webp" type="image/x-icon">
 
-    <title>Rhythm Messengers</title>
+    <title>Chicago Drumline | Rhythm Messengers</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
