+++ src/Video.js
@@ -1,4 +1,4 @@
-import React, { lazy } from "react";
+import React from "react";
 import { Grid, CardMedia } from "@mui/material";
 
 function Video() {
