+++ src/components/SEO.js
@@ -1,11 +1,9 @@
 import React from "react";
 import { Helmet } from "react-helmet-async";
 
-const SEO = ({ title, meta }) => {
+const SEO = ({ meta }) => {
   return (
     <Helmet>
-      {/* remove title */}
-      <title>{title}</title>
       {meta.length > 0 &&
         meta.map(({ name, content }) => {
           return <meta key={name} name={name} content={content}></meta>;
