+++ src/index.js
@@ -1,15 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import routes from "./routes";
 import "./index.css";
 import { HelmetProvider } from "react-helmet-async";
-
-const router = createBrowserRouter(routes);
+import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <HelmetProvider>
-    <RouterProvider router={router} />
+    <App />
   </HelmetProvider>
 );
